+++ src/common.rs
@@ -9,6 +9,7 @@ pub mod kernel_parameters_ops;
 pub mod local_git_repo;
 pub mod log_ops;
 pub mod node_ops;
+pub mod pcs_utils;
 pub mod sat_file;
 pub mod terminal_ops;
 pub mod vault;